Optimering

+++ Settings.h
@@ -29,12 +29,12 @@ public:
 	float altKd = 0.0;
 
 	// int16_t speedOff1 = 8, speedOff2 = 0, speedOff3 = 9, speedOff4 = 3;
-	int16_t offset1 = 0;
-	int16_t offset2 = 0;
-	int16_t offset3 = 0;
-	int16_t offset4 = 0;
+	int offset1 = 0;
+	int offset2 = 0;
+	int offset3 = 0;
+	int offset4 = 0;
 
-	int16_t hoverOffset = 0;
+	int hoverOffset = 0;
 
 public:
 	/* Settings functions */
